Make sure serde tests run in CI
