+++ .gitignore
@@ -8,3 +8,5 @@ book
 Cargo.lock
 
 petitset_serde_test/target/
+
+serde_tests/target/
