+++ serde_tests/Cargo.toml
@@ -2,7 +2,7 @@
 # https://github.com/rust-lang/cargo/issues/2911
 
 [package]
-name = "petitset_serde_test"
+name = "serde_tests"
 version = "0.0.0"
 edition = "2021"
